refactor: remove deprecated WriteAheadLog, add SAFETY docs and health monitoring
- Remove WriteAheadLog, AsyncWalWriter, and FlushGuard (deprecated by ConcurrentWalSystem)
- Add SegmentReader for WAL recovery operations
- Add comprehensive SAFETY comments for unsafe code in ring buffer
- Document memory ordering rationale for Relaxed LSN allocation
- Add flush error tracking with consecutive_flush_errors counter
- Add is_healthy() method for WAL health monitoring
- Update all benchmarks to use ConcurrentWalSystem
- Remove ~4300 lines of dead code
