+++ Cargo.toml
@@ -188,12 +188,6 @@ path = "benches/hnsw_index.rs"
 required-features = []
 
 [[bench]]
-name = "async_wal_writer"
-harness = false
-path = "benches/async_wal_writer.rs"
-required-features = []
-
-[[bench]]
 name = "wal_serialization"
 harness = false
 path = "benches/wal_serialization.rs"
