--- benches/async_wal_writer.rs
@@ -1,167 +1,0 @@
-//! Benchmarks for AsyncWalWriter performance.
-//!
-//! This benchmark validates that the AsyncWalWriter meets the performance targets:
-//! - Write latency: <10µs per operation (actual: ~118ns)
-//! - Throughput: >100,000 writes/sec (actual: ~6M writes/sec)
-//!
-//! # Test Conditions
-//!
-//! - **Buffer size**: 10,000-20,000 entries (configurable per test)
-//! - **Sync interval**: 10-100ms (configurable per test)
-//! - **Entry size**: ~200 bytes (CreateNode operation with minimal properties)
-//! - **Hardware**: Performance will vary by system (SSD vs HDD, CPU speed)
-//!
-//! # Performance Characteristics
-//!
-//! - **Latency**: append() is lock-free and returns in <200ns (channel send)
-//! - **Throughput**: Scales linearly with buffer size up to ~10M ops/sec
-//! - **Batching**: Automatically batches entries for efficient fsync
-//!
-//! # Scaling Guidelines
-//!
-//! - Larger buffer → higher throughput, more memory usage, longer data-at-risk window
-//! - Smaller buffer → lower latency to disk, less memory, more frequent fsync
-//! - Optimal buffer size depends on workload (typically 1000-10000)
-
-mod common;
-
-use criterion::{Criterion, Throughput, black_box, criterion_group, criterion_main};
-use gallifreydb::core::id::NodeId;
-use gallifreydb::core::property::PropertyMap;
-use gallifreydb::core::temporal::BiTemporalInterval;
-use gallifreydb::storage::wal::{AsyncWalWriter, LSN, WalEntry, WalOperation};
-use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::time::Duration;
-
-fn create_test_entry(lsn: u64) -> WalEntry {
-    WalEntry {
-        lsn: LSN(lsn),
-        timestamp: 0,
-        operation: WalOperation::CreateNode {
-            node_id: NodeId::new(lsn).expect("valid node ID"),
-            label: "BenchNode".to_string(),
-            properties: PropertyMap::new(),
-            temporal: BiTemporalInterval::current(0),
-        },
-        checksum: 0,
-    }
-}
-
-fn bench_async_write_latency(c: &mut Criterion) {
-    let mut group = c.benchmark_group("async_wal_writer");
-    group.throughput(Throughput::Elements(1));
-
-    group.bench_function("single_write_latency", |b| {
-        let write_count = Arc::new(AtomicU64::new(0));
-        let write_count_clone = Arc::clone(&write_count);
-
-        // Use a large buffer to avoid backpressure
-        let writer = AsyncWalWriter::new(
-            10000,
-            Duration::from_millis(100),
-            move |batch| {
-                write_count_clone.fetch_add(batch.len() as u64, Ordering::Relaxed);
-            },
-            vec![],
-        );
-
-        let mut lsn = 1;
-        b.iter(|| {
-            let entry = create_test_entry(lsn);
-            lsn += 1;
-            writer.append(black_box(entry)).unwrap();
-        });
-    });
-
-    group.finish();
-}
-
-fn bench_async_write_throughput(c: &mut Criterion) {
-    let mut group = c.benchmark_group("async_wal_writer");
-    group.throughput(Throughput::Elements(10000));
-    group.sample_size(50); // Reduce sample size for faster benchmarking
-
-    group.bench_function("throughput_10k_writes", |b| {
-        b.iter(|| {
-            let write_count = Arc::new(AtomicU64::new(0));
-            let write_count_clone = Arc::clone(&write_count);
-
-            // Large buffer to avoid backpressure
-            let writer = AsyncWalWriter::new(
-                20000,
-                Duration::from_millis(100),
-                move |batch| {
-                    write_count_clone.fetch_add(batch.len() as u64, Ordering::Relaxed);
-                },
-                vec![],
-            );
-
-            // Write 10,000 entries
-            for lsn in 1..=10_000 {
-                let entry = create_test_entry(lsn);
-                writer.append(black_box(entry)).unwrap();
-            }
-
-            // Drop writer to ensure all entries are flushed
-            drop(writer);
-        });
-    });
-
-    group.finish();
-}
-
-fn bench_async_write_batching(c: &mut Criterion) {
-    let mut group = c.benchmark_group("async_wal_writer");
-
-    group.bench_function("batch_efficiency", |b| {
-        b.iter(|| {
-            let batch_sizes = Arc::new(std::sync::Mutex::new(Vec::new()));
-            let batch_sizes_clone = Arc::clone(&batch_sizes);
-
-            let writer = AsyncWalWriter::new(
-                1000,
-                Duration::from_millis(10),
-                move |batch| {
-                    let mut sizes = batch_sizes_clone.lock().unwrap();
-                    sizes.push(batch.len());
-                },
-                vec![],
-            );
-
-            // Write entries rapidly to encourage batching
-            for lsn in 1..=1000 {
-                let entry = create_test_entry(lsn);
-                writer.append(black_box(entry)).unwrap();
-            }
-
-            // Wait for processing
-            drop(writer);
-
-            let sizes = batch_sizes.lock().unwrap();
-            let avg_batch_size: f64 = if !sizes.is_empty() {
-                sizes.iter().sum::<usize>() as f64 / sizes.len() as f64
-            } else {
-                0.0
-            };
-
-            // Print batch statistics for verification
-            println!(
-                "  Batches: {}, Avg size: {:.1}",
-                sizes.len(),
-                avg_batch_size
-            );
-        });
-    });
-
-    group.finish();
-}
-
-criterion_group!(
-    name = benches;
-    config = common::configure_criterion();
-    targets = bench_async_write_latency,
-    bench_async_write_throughput,
-    bench_async_write_batching
-);
-criterion_main!(benches);
