+++ src/storage/wal.rs
@@ -6,62 +6,70 @@
 //! - Point-in-time recovery capabilities
 //! - Configurable durability modes for performance tuning
 //!
-//! # WAL Format
+//! # Architecture
 //!
-//! Each WAL entry consists of:
-//! - Log Sequence Number (LSN): Monotonically increasing identifier
-//! - Timestamp: When the operation was logged
-//! - Operation: The mutation to apply
-//! - Checksum: CRC32 for corruption detection
+//! GallifreyDB uses a **Concurrent WAL with Striped Lock-Free Ring Buffers** for
+//! high-throughput write operations while maintaining ACID compliance.
+//!
+//! ```text
+//!                     ┌─────────────────────┐
+//!                     │    LSN Allocator    │
+//!                     │  AtomicU64::fetch_add
+//!                     └──────────┬──────────┘
+//!                                │
+//!        ┌───────────────────────┼───────────────────────┐
+//!        ▼                       ▼                       ▼
+//! ┌─────────────┐         ┌─────────────┐         ┌─────────────┐
+//! │   Stripe 0  │         │   Stripe 1  │         │  Stripe N   │
+//! │ Ring Buffer │         │ Ring Buffer │         │ Ring Buffer │
+//! │ (Lock-free) │         │ (Lock-free) │         │ (Lock-free) │
+//! └──────┬──────┘         └──────┬──────┘         └──────┬──────┘
+//!        └───────────────────────┼───────────────────────┘
+//!                                ▼
+//!                     ┌─────────────────────┐
+//!                     │  Flush Coordinator  │
+//!                     │  - Sorts by LSN     │
+//!                     │  - Writes segment   │
+//!                     └─────────────────────┘
+//! ```
+//!
+//! # Key Design Principles
+//!
+//! 1. **Lock-free append path**: Multiple threads can append concurrently without mutex contention
+//! 2. **Global LSN ordering**: Single atomic counter ensures total ordering of all operations
+//! 3. **Sorted flush**: Entries are sorted by LSN before writing to disk
+//! 4. **ACID preserved**: Synchronous and GroupCommit modes remain fully ACID compliant
 //!
 //! # Durability Modes
 //!
-//! The WAL supports three durability modes:
-//!
-//! - **Synchronous**: fsync after every commit (maximum durability, default)
-//! - **Async**: Background thread flushes periodically (maximum throughput)
-//! - **GroupCommit**: Batch multiple transactions per fsync (ACID + high throughput)
+//! | Mode | Latency | Throughput | ACID |
+//! |------|---------|------------|------|
+//! | Synchronous | ~1.5ms | ~600/sec | ✅ Full |
+//! | GroupCommit | ~10-50ms | ~100K+/sec | ✅ Full |
+//! | Async | <100ns | ~500K+/sec | ❌ Eventual |
 //!
 //! See [`DurabilityMode`] for details.
 //!
-//! # Async Mode Integration
+//! # Usage
 //!
-//! The Async durability mode is implemented via [`AsyncWalWriter`], which provides
-//! non-blocking WAL appends with background batched fsync operations.
+//! ```ignore
+//! use gallifreydb::storage::wal::concurrent_system::{ConcurrentWalSystem, ConcurrentWalSystemConfig};
 //!
-//! ## Lifecycle
+//! let config = ConcurrentWalSystemConfig::new("data/wal");
+//! let wal = ConcurrentWalSystem::new(config)?;
 //!
-//! 1. **Creation**: `AsyncWalWriter` is created when the WAL is opened with `DurabilityMode::Async`
-//! 2. **Operation**: Entries are appended to a bounded channel (default: 1000 entries)
-//! 3. **Background Thread**: Continuously drains channel and performs batched fsyncs
-//! 4. **Shutdown**: Drop impl signals background thread, drains remaining entries, and syncs
+//! // Async append (returns immediately)
+//! let lsn = wal.append_async(operation)?;
 //!
-//! ## Integration Points
+//! // Commit with configured durability
+//! wal.commit()?;
 //!
-//! - **append()**: Writes entries to BufWriter, then sends to AsyncWalWriter channel
-//! - **Piggyback Optimization**: Synchronous commits drain async entries to avoid redundant fsyncs
-//! - **Error Handling**: Background thread failures are detected and logged
-//! - **Recovery**: On restart, all WAL segments are replayed (no special async handling needed)
-//!
-//! ## Thread Safety
-//!
-//! The integration uses several synchronization primitives:
-//! - `crossbeam_channel` for lock-free communication
-//! - `Arc<AtomicBool>` for thread liveness detection
-//! - Control channel for drain commands (piggyback optimization)
-//! - All metrics use `Ordering::Relaxed` (observability only, not for coordination)
-//!
-//! ## Performance Characteristics
-//!
-//! - **append() latency**: ~118ns (channel send + metric update)
-//! - **Throughput**: ~6M writes/sec on modern hardware
-//! - **Batch efficiency**: 10-100 entries per fsync under normal load
-//! - **Backpressure**: Blocks caller when channel is full (bounded buffer)
+//! // Shutdown gracefully
+//! wal.shutdown();
+//! ```
 
-// Submodules for durability mode support
-pub mod async_writer;
+// Durability mode support
 pub mod durability;
-pub mod flush_guard;
 pub mod group_commit;
 
 // Concurrent WAL modules
@@ -70,12 +78,11 @@ pub mod concurrent_system;
 pub mod flush_coordinator;
 pub mod lsn_allocator;
 pub mod ring_buffer;
+pub mod segment_reader;
 pub mod stripe;
 
 // Re-export key types
-pub use async_writer::{AsyncWalMetrics, AsyncWalWriter, WalEvent, WalObserver};
 pub use durability::{DurabilityMode, WriteOptions};
-pub use flush_guard::{FlushGuard, FlushSignal};
 pub use group_commit::GroupCommitCoordinator;
 
 use crate::core::{
@@ -83,20 +90,7 @@ use crate::core::{
     property::PropertyMap,
     temporal::{BiTemporalInterval, Timestamp, time},
 };
-use crate::utils::error::{Error, Result, StorageError};
-use std::fs::{File, OpenOptions};
-use std::io::{BufWriter, Write};
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, Mutex};
-
-/// Magic bytes identifying a GallifreyDB WAL segment file.
-const WAL_MAGIC: [u8; 4] = *b"GWAL";
-
-/// Current WAL format version.
-const WAL_VERSION: u8 = 1;
-
-/// Size of the WAL segment header (magic + version).
-const WAL_HEADER_SIZE: usize = 5;
+use std::path::PathBuf;
 
 /// Log Sequence Number - monotonically increasing identifier for WAL entries
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -302,48 +296,11 @@ impl WalConfig {
     }
 }
 
-/// Initial capacity for the WAL serialization buffer (Issue #20)
-/// 2KB is sufficient for most entries without reallocation,
-/// while the buffer automatically grows for larger entries.
-const SERIALIZE_BUFFER_INITIAL_CAPACITY: usize = 2048;
+// =============================================================================
+// Serialization Helpers
+// =============================================================================
 
-/// Write-Ahead Log manager
-///
-/// # Lock-Free Sync Architecture
-///
-/// To achieve high throughput (50k+ ops/sec), the WAL uses a two-phase flush:
-///
-/// 1. **Append Phase** (fast, µs): Data written to `writer` (BufWriter), flushed to OS page cache
-/// 2. **Sync Phase** (slow, ms): `sync_handle` forces OS to write page cache to disk
-///
-/// The key insight is that `sync_handle` is a cloned file handle that can sync
-/// independently. This allows new writes to pile up in the OS page cache while
-/// a previous sync is in progress, using the OS page cache as a natural double-buffer.
-pub struct WriteAheadLog {
-    config: WalConfig,
-    current_lsn: LSN,
-    current_segment: u64,
-    /// The buffered writer for appending entries (fast path)
-    writer: Option<BufWriter<File>>,
-    /// Cloned file handle used exclusively for sync_data() calls.
-    /// This allows releasing the writer lock before the slow fsync,
-    /// enabling parallel writes during sync.
-    sync_handle: Option<File>,
-    current_size: usize,
-    /// Background flush thread guard (for Async and GroupCommit modes)
-    flush_guard: Option<FlushGuard>,
-    /// Signal for requesting immediate flush (for batch size triggering)
-    flush_signal: Option<Arc<FlushSignal>>,
-    /// Group commit coordinator (for GroupCommit mode)
-    group_commit: Option<Arc<GroupCommitCoordinator>>,
-    /// Async WAL writer (for Async durability mode with recv_timeout pattern)
-    async_writer: Option<AsyncWalWriter>,
-    /// Tracks whether there's unflushed data (for piggybacking optimization)
-    has_pending_data: bool,
-    /// Reusable serialization buffer to avoid allocating Vec per entry (Issue #20)
-    /// No Mutex needed - append() has &mut self, guaranteeing exclusive access
-    serialize_buffer: Vec<u8>,
-}
+use crate::utils::error::Result;
 
 /// Helper to serialize a string into the buffer (length prefix + bytes)
 #[inline(always)]
@@ -354,9 +311,9 @@ fn serialize_str(s: &str, buffer: &mut Vec<u8>) {
 
 /// Serialize a WAL entry with CRC32 checksum into the provided buffer
 ///
-/// This function reuses the provided buffer to avoid per-entry allocation (Issue #20).
+/// This function reuses the provided buffer to avoid per-entry allocation.
 /// The caller should clear the buffer before calling this function to maintain its capacity.
-fn serialize_entry_into(entry: &WalEntry, buffer: &mut Vec<u8>) -> Result<()> {
+pub(crate) fn serialize_entry_into(entry: &WalEntry, buffer: &mut Vec<u8>) -> Result<()> {
     // Write LSN (8 bytes)
     buffer.extend_from_slice(&entry.lsn.0.to_le_bytes());
 
@@ -452,2353 +409,4 @@ fn serialize_entry_into(entry: &WalEntry, buffer: &mut Vec<u8>) -> Result<()> {
     buffer[checksum_offset..checksum_offset + 4].copy_from_slice(&checksum.to_le_bytes());
 
     Ok(())
-}
-
-impl WriteAheadLog {
-    /// Create a new WAL with the given configuration
-    pub fn new(config: WalConfig) -> Result<Self> {
-        // Create WAL directory if it doesn't exist
-        std::fs::create_dir_all(&config.wal_dir)
-            .map_err(|e| StorageError::IoError(format!("Failed to create WAL directory: {}", e)))?;
-
-        let durability_mode = config.durability_mode;
-
-        let mut wal = WriteAheadLog {
-            config,
-            current_lsn: LSN::initial(),
-            current_segment: 1,
-            writer: None,
-            sync_handle: None,
-            current_size: 0,
-            flush_guard: None,
-            flush_signal: None,
-            group_commit: None,
-            async_writer: None,
-            has_pending_data: false,
-            serialize_buffer: Vec::with_capacity(SERIALIZE_BUFFER_INITIAL_CAPACITY),
-        };
-
-        // Initialize background infrastructure based on durability mode
-        wal.initialize_durability_mode(durability_mode)?;
-
-        Ok(wal)
-    }
-
-    /// Initialize the durability mode infrastructure.
-    ///
-    /// This sets up background threads and coordinators based on the mode.
-    fn initialize_durability_mode(&mut self, mode: DurabilityMode) -> Result<()> {
-        match mode {
-            DurabilityMode::Synchronous => {
-                // No background infrastructure needed
-                self.flush_guard = None;
-                self.group_commit = None;
-            }
-            DurabilityMode::Async { flush_interval_ms } => {
-                // Set up group commit coordinator (not used but simplifies piggybacking)
-                self.group_commit = None;
-
-                // Note: FlushGuard will be created when the WAL is wrapped in Arc<Mutex<>>
-                // and passed to GallifreyDB, since we need a reference to call flush_to_disk
-                // For now, store the interval in config and create guard later
-                let _ = flush_interval_ms; // Will be used when creating FlushGuard
-            }
-            DurabilityMode::GroupCommit {
-                max_delay_ms,
-                max_batch_size,
-            } => {
-                // Set up group commit coordinator
-                self.group_commit = Some(Arc::new(GroupCommitCoordinator::new(
-                    max_delay_ms,
-                    max_batch_size,
-                )));
-            }
-            DurabilityMode::AsyncBatched {
-                max_delay_ms,
-                max_batch_size,
-            } => {
-                // Set up group commit coordinator for epoch tracking (like GroupCommit)
-                // but transactions won't block on waiting
-                self.group_commit = Some(Arc::new(GroupCommitCoordinator::new(
-                    max_delay_ms,
-                    max_batch_size,
-                )));
-            }
-        }
-        Ok(())
-    }
-
-    /// Get the durability mode.
-    pub fn durability_mode(&self) -> DurabilityMode {
-        self.config.durability_mode
-    }
-
-    /// Get a reference to the group commit coordinator, if any.
-    pub fn group_commit_coordinator(&self) -> Option<&Arc<GroupCommitCoordinator>> {
-        self.group_commit.as_ref()
-    }
-
-    /// Check if there is pending unflushed data.
-    pub fn has_pending_data(&self) -> bool {
-        self.has_pending_data
-    }
-
-    /// Start the background flush thread for Async and GroupCommit modes.
-    ///
-    /// This must be called after the WAL is wrapped in `Arc<Mutex<>>` since
-    /// the flush thread needs a reference to call flush methods.
-    ///
-    /// # Arguments
-    ///
-    /// * `wal` - The Arc<Mutex<WAL>> that this WAL instance is wrapped in
-    pub fn start_flush_thread(wal: Arc<Mutex<Self>>) {
-        use std::fs::File;
-        use std::time::Duration;
-
-        // Get the durability mode and group commit coordinator
-        let (mode, group_commit): (DurabilityMode, Option<Arc<GroupCommitCoordinator>>) = {
-            let wal_guard = wal.lock().expect("WAL lock poisoned");
-            (
-                wal_guard.config.durability_mode,
-                wal_guard.group_commit.clone(),
-            )
-        };
-
-        let interval = match mode {
-            DurabilityMode::Synchronous => {
-                // No background thread needed for Synchronous mode
-                return;
-            }
-            DurabilityMode::Async { flush_interval_ms } => Duration::from_millis(flush_interval_ms),
-            DurabilityMode::GroupCommit { max_delay_ms, .. } => Duration::from_millis(max_delay_ms),
-            DurabilityMode::AsyncBatched { max_delay_ms, .. } => {
-                Duration::from_millis(max_delay_ms)
-            }
-        };
-
-        // Handle Async mode with AsyncWalWriter (recv_timeout pattern)
-        if matches!(mode, DurabilityMode::Async { .. }) {
-            let wal_clone = Arc::clone(&wal);
-            let flush_interval = interval;
-
-            // Create AsyncWalWriter with write_fn that flushes and syncs the WAL
-            let async_writer = AsyncWalWriter::new(
-                1000, // Buffer size for entries
-                flush_interval,
-                move |batch| {
-                    // Phase 1: Flush to OS cache (fast, with lock)
-                    let sync_handle_for_batch = {
-                        let mut wal_guard = match wal_clone.lock() {
-                            Ok(guard) => guard,
-                            Err(_) => return,
-                        };
-
-                        // Flush accumulated writes to OS cache
-                        if let Err(_e) = wal_guard.flush_to_os() {
-                            // Log error in production, but continue
-                            return;
-                        }
-
-                        // Clone sync handle for this batch
-                        wal_guard
-                            .sync_handle
-                            .as_ref()
-                            .and_then(|h| h.try_clone().ok())
-                    }; // Lock released - parallel writes can continue
-
-                    // Phase 2: Sync to disk (slow, no lock held)
-                    if let Some(ref sync_handle) = sync_handle_for_batch
-                        && let Err(_e) = sync_handle.sync_data()
-                    {
-                        // Log error in production
-                        return;
-                    }
-
-                    // Update metrics
-                    let _ = batch.len(); // Batch size available for metrics
-                },
-                vec![], // No observers for now
-            );
-
-            // Store AsyncWalWriter in the WAL
-            let mut wal_guard = wal.lock().expect("WAL lock poisoned");
-            wal_guard.async_writer = Some(async_writer);
-            return;
-        }
-
-        let wal_clone: Arc<Mutex<WriteAheadLog>> = Arc::clone(&wal);
-        let group_commit_clone: Option<Arc<GroupCommitCoordinator>> = group_commit.clone();
-
-        let flush_guard = FlushGuard::spawn(
-            move || {
-                // Phase 1: Flush to OS page cache and get current sync handle
-                // CRITICAL: We must get the sync handle on each iteration to handle segment rotation.
-                // If we held a pre-cloned handle, it would become stale after rotation and we'd
-                // be fsyncing the wrong (old/deleted) segment file!
-                let sync_handle_for_iteration: Option<File> = {
-                    let mut wal_guard: std::sync::MutexGuard<'_, WriteAheadLog> =
-                        match wal_clone.lock() {
-                            Ok(guard) => guard,
-                            Err(_) => return,
-                        };
-
-                    if let Err(_e) = wal_guard.flush_to_os() {
-                        // Log error in production, but continue
-                    }
-
-                    // Clone the current sync handle (handles segment rotation correctly)
-                    // NOTE: try_clone() on already-cloned FDs is cheap - reuses file description
-                    wal_guard
-                        .sync_handle
-                        .as_ref()
-                        .and_then(|h| h.try_clone().ok())
-                }; // Lock released - parallel writes can now pile up
-
-                // Phase 2: Sync to disk (slow, ms) - no lock held!
-                // NOTE: sync_handle may be None for fresh WAL instances before first write.
-                // This is normal - we skip the sync and let the flush complete successfully.
-                if let Some(ref sync_handle) = sync_handle_for_iteration
-                    && let Err(e) = sync_handle.sync_data()
-                {
-                    // Mark flush as failed if using GroupCommit
-                    if let Some(ref gc) = group_commit_clone {
-                        let err =
-                            Error::Storage(StorageError::IoError(format!("fsync failed: {}", e)));
-                        gc.mark_flushed(Err(err));
-                    }
-                    return;
-                }
-                // If no sync handle (fresh WAL), skip sync but continue to notify waiters
-
-                // Phase 3: Notify waiters
-                // NOTE: We do NOT reset has_pending_data here to avoid a race condition:
-                // - New writes may have occurred after we cloned the file handle (Phase 2)
-                // - Those writes may or may not have been flushed by our fsync
-                // - Setting the flag to false could incorrectly indicate no pending data
-                // The flag will be reset by synchronous flush paths that hold the lock throughout.
-
-                // Notify GroupCommit waiters that flush is complete
-                if let Some(ref gc) = group_commit_clone {
-                    gc.mark_flushed(Ok(()));
-                }
-
-                // NOTE: Background flush observability metrics removed to prevent log spam.
-                // See GitHub issue #274 for tracking proper observability implementation.
-            },
-            interval,
-        );
-
-        // Store the flush guard and signal in the WAL
-        let mut wal_guard = wal.lock().expect("WAL lock poisoned");
-        wal_guard.flush_signal = Some(Arc::clone(flush_guard.signal()));
-        wal_guard.flush_guard = Some(flush_guard);
-    }
-
-    /// Open a new WAL segment for writing
-    fn open_segment(&mut self, segment_id: u64) -> Result<()> {
-        let segment_path = self.segment_path(segment_id);
-
-        let file = OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(&segment_path)
-            .map_err(|e| StorageError::IoError(format!("Failed to open WAL segment: {}", e)))?;
-
-        // Clone the file handle for sync operations.
-        // This allows us to release the writer lock before calling sync_data(),
-        // enabling parallel writes during the slow fsync operation.
-        let sync_handle = file.try_clone().map_err(|e| {
-            StorageError::IoError(format!("Failed to clone WAL file handle: {}", e))
-        })?;
-
-        // Get current file size
-        let metadata = file.metadata().map_err(|e| {
-            StorageError::IoError(format!("Failed to get WAL segment metadata: {}", e))
-        })?;
-
-        let file_size = metadata.len() as usize;
-        // Use 64KB buffer (vs default 8KB) to reduce write syscall overhead and improve
-        // performance for all durability modes, especially Async which spends 1.8ms in
-        // BufWriter::flush(). This should reduce that to ~1ms.
-        let mut writer = BufWriter::with_capacity(64 * 1024, file);
-
-        // Write header for new segments
-        if file_size == 0 {
-            writer
-                .write_all(&WAL_MAGIC)
-                .map_err(|e| StorageError::IoError(format!("Failed to write WAL magic: {}", e)))?;
-            writer.write_all(&[WAL_VERSION]).map_err(|e| {
-                StorageError::IoError(format!("Failed to write WAL version: {}", e))
-            })?;
-            self.current_size = WAL_HEADER_SIZE;
-        } else {
-            self.current_size = file_size;
-        }
-
-        self.writer = Some(writer);
-        self.sync_handle = Some(sync_handle);
-
-        Ok(())
-    }
-
-    /// Get the path for a WAL segment
-    fn segment_path(&self, segment_id: u64) -> PathBuf {
-        self.config.wal_dir.join(format!("{:06}.log", segment_id))
-    }
-
-    /// Append an operation to the WAL
-    ///
-    /// This writes the operation to the BufWriter but does NOT fsync.
-    /// Call `flush_to_os()` to push to OS cache, then `sync_to_disk()` for durability.
-    pub fn append(&mut self, operation: WalOperation) -> Result<LSN> {
-        // Ensure we have a writer
-        if self.writer.is_none() {
-            self.open_segment(self.current_segment)?;
-        }
-
-        let entry = WalEntry::new(self.current_lsn, operation);
-
-        // Serialize using the reusable buffer (zero-allocation after warmup)
-        self.serialize_buffer.clear();
-        serialize_entry_into(&entry, &mut self.serialize_buffer)?;
-
-        let serialized_len = self.serialize_buffer.len();
-
-        // Write directly from the reused buffer
-        if let Some(writer) = &mut self.writer {
-            writer
-                .write_all(&self.serialize_buffer)
-                .map_err(|e| StorageError::IoError(format!("Failed to write WAL entry: {}", e)))?;
-        }
-
-        // Mark that we have pending unflushed data
-        self.has_pending_data = true;
-        self.current_size += serialized_len;
-
-        let lsn = self.current_lsn;
-        self.current_lsn = self.current_lsn.next();
-
-        // For Async mode, send entry to AsyncWalWriter for batched syncing
-        // Note: We don't flush_to_os() here - the background thread handles flushing
-        // This keeps append() fast and non-blocking
-        if let Some(ref async_writer) = self.async_writer
-            && let Err(_e) = async_writer.append(entry.clone())
-        {
-            // WAL background sync thread has terminated unexpectedly - this is an error condition
-            // but we've already written to the BufWriter, so continue
-            #[cfg(feature = "observability")]
-            tracing::error!(
-                error = ?_e,
-                lsn = ?entry.lsn,
-                "AsyncWalWriter background thread terminated unexpectedly"
-            );
-        }
-
-        // Check if we need to rotate to a new segment
-        if self.current_size >= self.config.segment_size {
-            self.rotate_segment()?;
-        }
-
-        Ok(lsn)
-    }
-
-    /// Rotate to a new WAL segment
-    fn rotate_segment(&mut self) -> Result<()> {
-        // Flush and sync current segment before rotation
-        if let Some(mut writer) = self.writer.take() {
-            writer.flush().map_err(|e| {
-                StorageError::IoError(format!("Failed to flush WAL on rotation: {}", e))
-            })?;
-        }
-
-        // Sync the old segment to disk
-        if let Some(sync_handle) = self.sync_handle.take() {
-            sync_handle.sync_data().map_err(|e| {
-                StorageError::IoError(format!("Failed to sync WAL on rotation: {}", e))
-            })?;
-        }
-
-        self.has_pending_data = false;
-
-        // Move to next segment
-        self.current_segment += 1;
-        self.current_size = 0;
-        self.open_segment(self.current_segment)?;
-
-        // Clean up old segments
-        self.cleanup_old_segments()?;
-
-        Ok(())
-    }
-
-    /// Remove old WAL segments beyond retention policy
-    fn cleanup_old_segments(&mut self) -> Result<()> {
-        let keep_from = self
-            .current_segment
-            .saturating_sub(self.config.segments_to_retain as u64);
-
-        if let Ok(entries) = std::fs::read_dir(&self.config.wal_dir) {
-            for entry in entries.flatten() {
-                if let Some(name) = entry.file_name().to_str()
-                    && name.ends_with(".log")
-                    && let Some(seg_id) = name
-                        .strip_suffix(".log")
-                        .and_then(|s| s.parse::<u64>().ok())
-                    && seg_id < keep_from
-                {
-                    let _ = std::fs::remove_file(entry.path());
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Read all WAL entries from a specific LSN onwards
-    pub fn read_from(&self, start_lsn: LSN) -> Result<Vec<WalEntry>> {
-        let mut entries = Vec::new();
-
-        // Find all WAL segments
-        let mut segments = Vec::new();
-        if let Ok(dir_entries) = std::fs::read_dir(&self.config.wal_dir) {
-            for entry in dir_entries.flatten() {
-                if let Some(name) = entry.file_name().to_str()
-                    && name.ends_with(".log")
-                    && let Some(seg_id) = name
-                        .strip_suffix(".log")
-                        .and_then(|s| s.parse::<u64>().ok())
-                {
-                    segments.push((seg_id, entry.path()));
-                }
-            }
-        }
-
-        // Sort segments by ID
-        segments.sort_by_key(|(id, _)| *id);
-
-        // Read entries from each segment
-        for (_, path) in segments {
-            let segment_entries = self.read_segment(&path, start_lsn)?;
-            entries.extend(segment_entries);
-        }
-
-        Ok(entries)
-    }
-
-    /// Helper to deserialize and validate a NodeId from WAL buffer
-    #[inline]
-    fn deserialize_node_id(buffer: &[u8], offset: usize, context: &str) -> Result<NodeId> {
-        let raw_id = u64::from_le_bytes([
-            buffer[offset],
-            buffer[offset + 1],
-            buffer[offset + 2],
-            buffer[offset + 3],
-            buffer[offset + 4],
-            buffer[offset + 5],
-            buffer[offset + 6],
-            buffer[offset + 7],
-        ]);
-        NodeId::new(raw_id).map_err(|e| {
-            Error::Storage(StorageError::CorruptedData(format!(
-                "Invalid node ID in WAL {}: {}",
-                context, e
-            )))
-        })
-    }
-
-    /// Helper to deserialize and validate an EdgeId from WAL buffer
-    #[inline]
-    fn deserialize_edge_id(buffer: &[u8], offset: usize, context: &str) -> Result<EdgeId> {
-        let raw_id = u64::from_le_bytes([
-            buffer[offset],
-            buffer[offset + 1],
-            buffer[offset + 2],
-            buffer[offset + 3],
-            buffer[offset + 4],
-            buffer[offset + 5],
-            buffer[offset + 6],
-            buffer[offset + 7],
-        ]);
-        EdgeId::new(raw_id).map_err(|e| {
-            Error::Storage(StorageError::CorruptedData(format!(
-                "Invalid edge ID in WAL {}: {}",
-                context, e
-            )))
-        })
-    }
-
-    /// Helper to deserialize and validate a VersionId from WAL buffer
-    #[inline]
-    fn deserialize_version_id(buffer: &[u8], offset: usize, context: &str) -> Result<VersionId> {
-        let raw_id = u64::from_le_bytes([
-            buffer[offset],
-            buffer[offset + 1],
-            buffer[offset + 2],
-            buffer[offset + 3],
-            buffer[offset + 4],
-            buffer[offset + 5],
-            buffer[offset + 6],
-            buffer[offset + 7],
-        ]);
-        VersionId::new(raw_id).map_err(|e| {
-            Error::Storage(StorageError::CorruptedData(format!(
-                "Invalid version ID in WAL {}: {}",
-                context, e
-            )))
-        })
-    }
-
-    /// Read WAL entries from a single segment file
-    fn read_segment(&self, path: &Path, start_lsn: LSN) -> Result<Vec<WalEntry>> {
-        use std::io::Read;
-
-        let file = match File::open(path) {
-            Ok(f) => f,
-            Err(_) => return Ok(Vec::new()), // File doesn't exist yet, return empty
-        };
-
-        let mut reader = std::io::BufReader::new(file);
-        let mut entries = Vec::new();
-        let mut buffer = Vec::new();
-
-        // Read entire file into buffer
-        reader
-            .read_to_end(&mut buffer)
-            .map_err(|e| StorageError::IoError(format!("Failed to read WAL segment: {}", e)))?;
-
-        // Detect WAL format version
-        let (version, mut offset) = if buffer.len() >= WAL_HEADER_SIZE && buffer[0..4] == WAL_MAGIC
-        {
-            // Version 1+ format: has magic header
-            let ver = buffer[4];
-            if ver > WAL_VERSION {
-                return Err(StorageError::CorruptedData(format!(
-                    "Unsupported WAL version: {} (max supported: {})",
-                    ver, WAL_VERSION
-                ))
-                .into());
-            }
-            (ver, WAL_HEADER_SIZE)
-        } else if !buffer.is_empty() {
-            // Invalid format: no magic header
-            return Err(StorageError::CorruptedData(
-                "Invalid WAL segment: missing GWAL magic header".to_string(),
-            )
-            .into());
-        } else {
-            return Ok(Vec::new()); // Empty segment
-        };
-        while offset < buffer.len() {
-            // Need at least 20 bytes for LSN (8) + timestamp (8) + checksum (4)
-            if offset + 20 > buffer.len() {
-                break;
-            }
-
-            // Read LSN (8 bytes)
-            let lsn = LSN(u64::from_le_bytes([
-                buffer[offset],
-                buffer[offset + 1],
-                buffer[offset + 2],
-                buffer[offset + 3],
-                buffer[offset + 4],
-                buffer[offset + 5],
-                buffer[offset + 6],
-                buffer[offset + 7],
-            ]));
-            offset += 8;
-
-            // Read timestamp (8 bytes)
-            let timestamp = i64::from_le_bytes([
-                buffer[offset],
-                buffer[offset + 1],
-                buffer[offset + 2],
-                buffer[offset + 3],
-                buffer[offset + 4],
-                buffer[offset + 5],
-                buffer[offset + 6],
-                buffer[offset + 7],
-            ]);
-            offset += 8;
-
-            // Read checksum (4 bytes)
-            let checksum = u32::from_le_bytes([
-                buffer[offset],
-                buffer[offset + 1],
-                buffer[offset + 2],
-                buffer[offset + 3],
-            ]);
-            offset += 4;
-
-            // Read operation type
-            if offset >= buffer.len() {
-                break;
-            }
-            let op_type = buffer[offset];
-            offset += 1;
-
-            // Parse operation data based on type and version
-            let operation = match op_type {
-                1 => {
-                    // CreateNode
-                    if offset + 12 > buffer.len() {
-                        break;
-                    }
-                    let node_id = Self::deserialize_node_id(&buffer, offset, "CreateNode")?;
-                    offset += 8;
-
-                    let label_len = u32::from_le_bytes([
-                        buffer[offset],
-                        buffer[offset + 1],
-                        buffer[offset + 2],
-                        buffer[offset + 3],
-                    ]) as usize;
-                    offset += 4;
-
-                    if offset + label_len > buffer.len() {
-                        break;
-                    }
-                    let label =
-                        String::from_utf8_lossy(&buffer[offset..offset + label_len]).to_string();
-                    offset += label_len;
-
-                    // V2+: deserialize properties and temporal
-                    // V1: use placeholders (data was not serialized)
-                    let (properties, temporal) = if version >= WAL_VERSION {
-                        let (props, props_len) = PropertyMap::deserialize(&buffer[offset..])?;
-                        offset += props_len;
-                        let (temp, temp_len) = BiTemporalInterval::deserialize(&buffer[offset..])?;
-                        offset += temp_len;
-                        (props, temp)
-                    } else {
-                        (PropertyMap::new(), BiTemporalInterval::current(timestamp))
-                    };
-
-                    WalOperation::CreateNode {
-                        node_id,
-                        label,
-                        properties,
-                        temporal,
-                    }
-                }
-                2 => {
-                    // CreateEdge
-                    if offset + 28 > buffer.len() {
-                        break;
-                    }
-                    let edge_id = Self::deserialize_edge_id(&buffer, offset, "CreateEdge")?;
-                    offset += 8;
-
-                    let source = Self::deserialize_node_id(&buffer, offset, "CreateEdge source")?;
-                    offset += 8;
-
-                    let target = Self::deserialize_node_id(&buffer, offset, "CreateEdge target")?;
-                    offset += 8;
-
-                    let label_len = u32::from_le_bytes([
-                        buffer[offset],
-                        buffer[offset + 1],
-                        buffer[offset + 2],
-                        buffer[offset + 3],
-                    ]) as usize;
-                    offset += 4;
-
-                    if offset + label_len > buffer.len() {
-                        break;
-                    }
-                    let label =
-                        String::from_utf8_lossy(&buffer[offset..offset + label_len]).to_string();
-                    offset += label_len;
-
-                    let (properties, temporal) = if version >= WAL_VERSION {
-                        let (props, props_len) = PropertyMap::deserialize(&buffer[offset..])?;
-                        offset += props_len;
-                        let (temp, temp_len) = BiTemporalInterval::deserialize(&buffer[offset..])?;
-                        offset += temp_len;
-                        (props, temp)
-                    } else {
-                        (PropertyMap::new(), BiTemporalInterval::current(timestamp))
-                    };
-
-                    WalOperation::CreateEdge {
-                        edge_id,
-                        source,
-                        target,
-                        label,
-                        properties,
-                        temporal,
-                    }
-                }
-                3 => {
-                    // UpdateNode
-                    // V1: only node_id + version_id (16 bytes)
-                    // V2: + label + properties + temporal
-                    if offset + 16 > buffer.len() {
-                        break;
-                    }
-                    let node_id = Self::deserialize_node_id(&buffer, offset, "UpdateNode")?;
-                    offset += 8;
-
-                    let version_id = Self::deserialize_version_id(&buffer, offset, "UpdateNode")?;
-                    offset += 8;
-
-                    let (label, properties, temporal) = if version >= WAL_VERSION {
-                        let label_len = u32::from_le_bytes([
-                            buffer[offset],
-                            buffer[offset + 1],
-                            buffer[offset + 2],
-                            buffer[offset + 3],
-                        ]) as usize;
-                        offset += 4;
-
-                        if offset + label_len > buffer.len() {
-                            break;
-                        }
-                        let lbl = String::from_utf8_lossy(&buffer[offset..offset + label_len])
-                            .to_string();
-                        offset += label_len;
-
-                        let (props, props_len) = PropertyMap::deserialize(&buffer[offset..])?;
-                        offset += props_len;
-                        let (temp, temp_len) = BiTemporalInterval::deserialize(&buffer[offset..])?;
-                        offset += temp_len;
-                        (lbl, props, temp)
-                    } else {
-                        (
-                            String::new(),
-                            PropertyMap::new(),
-                            BiTemporalInterval::current(timestamp),
-                        )
-                    };
-
-                    WalOperation::UpdateNode {
-                        node_id,
-                        version_id,
-                        label,
-                        properties,
-                        temporal,
-                    }
-                }
-                4 => {
-                    // UpdateEdge
-                    // V1: only edge_id + version_id (16 bytes)
-                    // V2: + label + properties + temporal
-                    if offset + 16 > buffer.len() {
-                        break;
-                    }
-                    let edge_id = Self::deserialize_edge_id(&buffer, offset, "UpdateEdge")?;
-                    offset += 8;
-
-                    let version_id = Self::deserialize_version_id(&buffer, offset, "UpdateEdge")?;
-                    offset += 8;
-
-                    let (label, properties, temporal) = if version >= WAL_VERSION {
-                        let label_len = u32::from_le_bytes([
-                            buffer[offset],
-                            buffer[offset + 1],
-                            buffer[offset + 2],
-                            buffer[offset + 3],
-                        ]) as usize;
-                        offset += 4;
-
-                        if offset + label_len > buffer.len() {
-                            break;
-                        }
-                        let lbl = String::from_utf8_lossy(&buffer[offset..offset + label_len])
-                            .to_string();
-                        offset += label_len;
-
-                        let (props, props_len) = PropertyMap::deserialize(&buffer[offset..])?;
-                        offset += props_len;
-                        let (temp, temp_len) = BiTemporalInterval::deserialize(&buffer[offset..])?;
-                        offset += temp_len;
-                        (lbl, props, temp)
-                    } else {
-                        (
-                            String::new(),
-                            PropertyMap::new(),
-                            BiTemporalInterval::current(timestamp),
-                        )
-                    };
-
-                    WalOperation::UpdateEdge {
-                        edge_id,
-                        version_id,
-                        label,
-                        properties,
-                        temporal,
-                    }
-                }
-                5 => {
-                    // Checkpoint (same format in all versions)
-                    if offset + 16 > buffer.len() {
-                        break;
-                    }
-                    let cp_lsn = LSN(u64::from_le_bytes([
-                        buffer[offset],
-                        buffer[offset + 1],
-                        buffer[offset + 2],
-                        buffer[offset + 3],
-                        buffer[offset + 4],
-                        buffer[offset + 5],
-                        buffer[offset + 6],
-                        buffer[offset + 7],
-                    ]));
-                    offset += 8;
-
-                    let cp_timestamp = i64::from_le_bytes([
-                        buffer[offset],
-                        buffer[offset + 1],
-                        buffer[offset + 2],
-                        buffer[offset + 3],
-                        buffer[offset + 4],
-                        buffer[offset + 5],
-                        buffer[offset + 6],
-                        buffer[offset + 7],
-                    ]);
-                    offset += 8;
-
-                    WalOperation::Checkpoint {
-                        lsn: cp_lsn,
-                        timestamp: cp_timestamp,
-                    }
-                }
-                6 => {
-                    // DeleteNode
-                    if offset + 8 > buffer.len() {
-                        break;
-                    }
-                    let node_id = Self::deserialize_node_id(&buffer, offset, "DeleteNode")?;
-                    offset += 8;
-
-                    let temporal = if version >= WAL_VERSION {
-                        let (temp, temp_len) = BiTemporalInterval::deserialize(&buffer[offset..])?;
-                        offset += temp_len;
-                        temp
-                    } else {
-                        BiTemporalInterval::current(timestamp)
-                    };
-
-                    WalOperation::DeleteNode { node_id, temporal }
-                }
-                7 => {
-                    // DeleteEdge
-                    if offset + 8 > buffer.len() {
-                        break;
-                    }
-                    let edge_id = Self::deserialize_edge_id(&buffer, offset, "DeleteEdge")?;
-                    offset += 8;
-
-                    let temporal = if version >= WAL_VERSION {
-                        let (temp, temp_len) = BiTemporalInterval::deserialize(&buffer[offset..])?;
-                        offset += temp_len;
-                        temp
-                    } else {
-                        BiTemporalInterval::current(timestamp)
-                    };
-
-                    WalOperation::DeleteEdge { edge_id, temporal }
-                }
-                _ => {
-                    // Unknown operation type
-                    return Err(StorageError::CorruptedData(format!(
-                        "Unknown WAL operation type: {}",
-                        op_type
-                    ))
-                    .into());
-                }
-            };
-
-            // Only include entries >= start_lsn
-            if lsn >= start_lsn {
-                let entry = WalEntry {
-                    lsn,
-                    timestamp,
-                    operation,
-                    checksum,
-                };
-                entries.push(entry);
-            }
-        }
-
-        Ok(entries)
-    }
-
-    /// Flush data to OS page cache (Phase 1 of two-phase flush).
-    ///
-    /// This is fast (microseconds) and should be done while holding the WAL lock.
-    /// After this call, data is in the OS page cache but NOT yet durable on disk.
-    ///
-    /// Call `sync_to_disk()` afterward (without holding the lock) for durability.
-    pub fn flush_to_os(&mut self) -> Result<()> {
-        #[cfg(feature = "observability")]
-        let _span = tracing::info_span!("wal_flush_to_os").entered();
-
-        if let Some(writer) = &mut self.writer {
-            writer
-                .flush()
-                .map_err(|e| StorageError::IoError(format!("Failed to flush WAL buffer: {}", e)))?;
-        }
-        Ok(())
-    }
-
-    /// Sync data from OS page cache to disk (Phase 2 of two-phase flush).
-    ///
-    /// This is slow (milliseconds) and should be done WITHOUT holding the WAL lock.
-    /// This allows parallel writes to pile up in the OS page cache during the fsync.
-    ///
-    /// Returns a reference to the sync handle for external callers who need to
-    /// sync outside the lock.
-    pub fn get_sync_handle(&self) -> Option<&File> {
-        self.sync_handle.as_ref()
-    }
-
-    /// Perform the slow disk sync using the sync handle.
-    ///
-    /// This should be called OUTSIDE the WAL lock to allow parallel writes.
-    pub fn sync_to_disk(&mut self) -> Result<()> {
-        if let Some(sync_handle) = &self.sync_handle {
-            sync_handle
-                .sync_data()
-                .map_err(|e| StorageError::IoError(format!("Failed to fsync WAL: {}", e)))?;
-        }
-        self.has_pending_data = false;
-        Ok(())
-    }
-
-    /// Full flush for backward compatibility (combines both phases).
-    ///
-    /// **Warning**: This holds the lock during fsync, limiting throughput.
-    /// For high-throughput scenarios, use `flush_to_os()` then release the lock,
-    /// then call `sync_to_disk()`.
-    pub fn flush(&mut self) -> Result<()> {
-        #[cfg(feature = "observability")]
-        let _span = tracing::info_span!("wal_flush_full").entered();
-
-        if let Some(writer) = &mut self.writer {
-            // Step 1: Flush BufWriter buffer to OS (fast)
-            {
-                #[cfg(feature = "observability")]
-                let _buffer_span = tracing::info_span!("wal_flush_buffer").entered();
-
-                writer.flush().map_err(|e| {
-                    StorageError::IoError(format!("Failed to flush WAL buffer: {}", e))
-                })?;
-            }
-
-            // Step 2: Force OS to sync data to disk (slow)
-            // SAFETY: sync_data() ensures durability by forcing the OS to write buffered data to disk.
-            // This is critical for WAL correctness - without it, committed transactions could be lost
-            // on crash/power failure. We use sync_data() instead of sync_all() because we only need
-            // to sync file data, not metadata (faster).
-            //
-            // NOTE: For high-throughput, call get_sync_handle() and sync outside the lock instead.
-            if let Some(sync_handle) = &self.sync_handle {
-                #[cfg(feature = "observability")]
-                let _fsync_span = tracing::info_span!("wal_fsync").entered();
-
-                sync_handle
-                    .sync_data()
-                    .map_err(|e| StorageError::IoError(format!("Failed to fsync WAL: {}", e)))?;
-            }
-        }
-        self.has_pending_data = false;
-        Ok(())
-    }
-
-    /// Commit with the specified durability mode.
-    ///
-    /// This is the main entry point for mode-aware transaction commits.
-    ///
-    /// # Returns
-    ///
-    /// - For Synchronous: Returns `None` (already durable)
-    /// - For Async: Returns `None` (will be flushed by background thread)
-    /// - For GroupCommit: Returns `Some(epoch)` to wait for
-    /// - For AsyncBatched: Returns `Some(epoch)` for optional tracking (caller doesn't block)
-    pub fn commit_with_mode(&mut self, mode: DurabilityMode) -> Result<Option<u64>> {
-        match mode {
-            DurabilityMode::Synchronous => {
-                // Piggyback optimization: Drain async writer entries before fsync
-                // This avoids redundant fsync by the background thread, since these
-                // entries are already in the BufWriter and will be synced by our flush()
-                if let Some(ref async_writer) = self.async_writer {
-                    // Use synchronous drain with acknowledgment to eliminate race conditions
-                    // If drain fails or times out (1ms), entries will still be synced by flush()
-                    let _ = async_writer.drain_without_sync_sync();
-                }
-
-                // Piggyback optimization: Wake background thread for GroupCommit
-                if let Some(signal) = &self.flush_signal {
-                    signal.request_flush();
-                }
-
-                // Immediate full flush (syncs both sync and drained async entries)
-                self.flush()?;
-                Ok(None)
-            }
-            DurabilityMode::Async { .. } => {
-                // Just flush to OS cache, background thread will sync
-                self.flush_to_os()?;
-                Ok(None)
-            }
-            DurabilityMode::GroupCommit { .. } => {
-                // Flush to OS cache and register with coordinator
-                self.flush_to_os()?;
-
-                if let Some(coordinator) = &self.group_commit {
-                    let (epoch, should_trigger) = coordinator.register_transaction();
-
-                    // If batch is full, wake the flush thread immediately
-                    if should_trigger && let Some(signal) = &self.flush_signal {
-                        signal.request_flush();
-                    }
-
-                    Ok(Some(epoch))
-                } else {
-                    // Fallback to sync if no coordinator (shouldn't happen)
-                    self.sync_to_disk()?;
-                    Ok(None)
-                }
-            }
-            DurabilityMode::AsyncBatched { .. } => {
-                // Phase 1: Flush to OS cache (fast, <100µs)
-                self.flush_to_os()?;
-
-                if let Some(coordinator) = &self.group_commit {
-                    // Register with coordinator but DON'T WAIT (key difference from GroupCommit)
-                    // The transaction returns immediately, but the epoch can be used for
-                    // optional durability tracking if needed.
-                    let (epoch, should_trigger) = coordinator.register_transaction();
-
-                    // Wake flush thread if batch is full
-                    if should_trigger && let Some(signal) = &self.flush_signal {
-                        signal.request_flush();
-                    }
-
-                    // Return epoch for optional tracking, but transaction doesn't block!
-                    Ok(Some(epoch))
-                } else {
-                    // Fallback: immediate sync (shouldn't happen in normal operation)
-                    self.sync_to_disk()?;
-                    Ok(None)
-                }
-            }
-        }
-    }
-
-    /// Wait for a group commit epoch to be flushed.
-    ///
-    /// This should be called after `commit_with_mode` returns `Some(epoch)`.
-    pub fn wait_for_epoch(&self, epoch: u64) -> Result<()> {
-        if let Some(coordinator) = &self.group_commit {
-            coordinator.wait_for_flush(epoch)?;
-        }
-        Ok(())
-    }
-
-    /// Get the current LSN
-    pub fn current_lsn(&self) -> LSN {
-        self.current_lsn
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_lsn_ordering() {
-        let lsn1 = LSN::initial();
-        let lsn2 = lsn1.next();
-        let lsn3 = lsn2.next();
-
-        assert!(lsn1 < lsn2);
-        assert!(lsn2 < lsn3);
-        assert_eq!(lsn1.0, 1);
-        assert_eq!(lsn2.0, 2);
-        assert_eq!(lsn3.0, 3);
-    }
-
-    #[test]
-    fn test_wal_entry_checksum() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let _wal = WriteAheadLog::new(config)?;
-
-        let operation = WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: "Person".to_string(),
-            properties: PropertyMap::new(),
-            temporal: BiTemporalInterval::current(time::now()),
-        };
-
-        let entry = WalEntry::new(LSN::initial(), operation);
-
-        // Serialize the entry to get the actual bytes with checksum
-        let mut serialized = Vec::new();
-        serialize_entry_into(&entry, &mut serialized)?;
-
-        // Verify checksum is valid
-        assert!(entry.verify_checksum(&serialized));
-
-        // Corrupt the checksum and verify it fails
-        let mut corrupted = serialized.clone();
-        corrupted[16] ^= 0xFF; // Flip some bits in the checksum
-        assert!(!entry.verify_checksum(&corrupted));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_creation() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let wal = WriteAheadLog::new(config)?;
-        assert_eq!(wal.current_lsn(), LSN::initial());
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_append() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        let operation = WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: "Person".to_string(),
-            properties: PropertyMap::new(),
-            temporal: BiTemporalInterval::current(time::now()),
-        };
-
-        let lsn = wal.append(operation)?;
-        assert_eq!(lsn, LSN::initial());
-        assert_eq!(wal.current_lsn(), LSN(2));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_segment_rotation() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            segment_size: 100, // Very small for testing rotation
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Append multiple entries to trigger rotation
-        for i in 0..10 {
-            let operation = WalOperation::CreateNode {
-                node_id: NodeId::new(i).unwrap(),
-                label: "Person".to_string(),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            };
-            wal.append(operation)?;
-        }
-
-        // Should have rotated to multiple segments
-        assert!(wal.current_segment > 1);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_delete_node_operation() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Log a delete node operation
-        let operation = WalOperation::DeleteNode {
-            node_id: NodeId::new(42).unwrap(),
-            temporal: BiTemporalInterval::current(time::now()),
-        };
-
-        let lsn = wal.append(operation)?;
-        assert_eq!(lsn, LSN::initial());
-
-        // Flush to ensure it's written
-        wal.flush()?;
-
-        // Read it back
-        let entries = wal.read_from(LSN::initial())?;
-        assert_eq!(entries.len(), 1);
-
-        match &entries[0].operation {
-            WalOperation::DeleteNode { node_id, .. } => {
-                assert_eq!(node_id.as_u64(), 42);
-            }
-            _ => panic!("Expected DeleteNode operation"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_delete_edge_operation() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Log a delete edge operation
-        let operation = WalOperation::DeleteEdge {
-            edge_id: EdgeId::new(99).unwrap(),
-            temporal: BiTemporalInterval::current(time::now()),
-        };
-
-        let lsn = wal.append(operation)?;
-        assert_eq!(lsn, LSN::initial());
-
-        // Flush to ensure it's written
-        wal.flush()?;
-
-        // Read it back
-        let entries = wal.read_from(LSN::initial())?;
-        assert_eq!(entries.len(), 1);
-
-        match &entries[0].operation {
-            WalOperation::DeleteEdge { edge_id, .. } => {
-                assert_eq!(edge_id.as_u64(), 99);
-            }
-            _ => panic!("Expected DeleteEdge operation"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_mixed_operations() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Log a sequence of mixed operations
-        let ops = vec![
-            WalOperation::CreateNode {
-                node_id: NodeId::new(1).unwrap(),
-                label: "Person".to_string(),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            },
-            WalOperation::CreateEdge {
-                edge_id: EdgeId::new(1).unwrap(),
-                source: NodeId::new(1).unwrap(),
-                target: NodeId::new(2).unwrap(),
-                label: "KNOWS".to_string(),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            },
-            WalOperation::DeleteNode {
-                node_id: NodeId::new(3).unwrap(),
-                temporal: BiTemporalInterval::current(time::now()),
-            },
-            WalOperation::DeleteEdge {
-                edge_id: EdgeId::new(2).unwrap(),
-                temporal: BiTemporalInterval::current(time::now()),
-            },
-        ];
-
-        for op in ops {
-            wal.append(op)?;
-        }
-
-        wal.flush()?;
-
-        // Read all entries back
-        let entries = wal.read_from(LSN::initial())?;
-        assert_eq!(entries.len(), 4);
-
-        // Verify operation types
-        assert!(matches!(
-            entries[0].operation,
-            WalOperation::CreateNode { .. }
-        ));
-        assert!(matches!(
-            entries[1].operation,
-            WalOperation::CreateEdge { .. }
-        ));
-        assert!(matches!(
-            entries[2].operation,
-            WalOperation::DeleteNode { .. }
-        ));
-        assert!(matches!(
-            entries[3].operation,
-            WalOperation::DeleteEdge { .. }
-        ));
-
-        Ok(())
-    }
-
-    // Tests for full serialization with properties and temporal data
-
-    #[test]
-    fn test_wal_create_node_full_roundtrip() -> Result<()> {
-        use crate::core::property::PropertyMapBuilder;
-        use crate::core::temporal::TimeRange;
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Create properties with various types
-        let properties = PropertyMapBuilder::new()
-            .insert("name", "Alice")
-            .insert("age", 30i64)
-            .insert("active", true)
-            .insert("score", 98.5f64)
-            .build();
-
-        // Create specific temporal interval
-        let temporal = BiTemporalInterval::new(
-            TimeRange::new(1000, 2000).unwrap(),
-            TimeRange::new(3000, crate::core::temporal::TIMESTAMP_MAX).unwrap(),
-        );
-
-        let operation = WalOperation::CreateNode {
-            node_id: NodeId::new(42).unwrap(),
-            label: "Person".to_string(),
-            properties: properties.clone(),
-            temporal,
-        };
-
-        wal.append(operation)?;
-        wal.flush()?;
-
-        // Read it back
-        let entries = wal.read_from(LSN::initial())?;
-        assert_eq!(entries.len(), 1);
-
-        match &entries[0].operation {
-            WalOperation::CreateNode {
-                node_id,
-                label,
-                properties: p,
-                temporal: t,
-            } => {
-                assert_eq!(node_id.as_u64(), 42);
-                assert_eq!(label, "Person");
-
-                // Verify properties
-                assert_eq!(p.get("name").and_then(|v| v.as_str()), Some("Alice"));
-                assert_eq!(p.get("age").and_then(|v| v.as_int()), Some(30));
-                assert_eq!(p.get("active").and_then(|v| v.as_bool()), Some(true));
-
-                // Verify temporal
-                assert_eq!(t.valid_time().start(), 1000);
-                assert_eq!(t.valid_time().end(), 2000);
-                assert_eq!(t.transaction_time().start(), 3000);
-                assert!(t.transaction_time().is_current());
-            }
-            _ => panic!("Expected CreateNode operation"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_create_edge_full_roundtrip() -> Result<()> {
-        use crate::core::property::PropertyMapBuilder;
-        use crate::core::temporal::TimeRange;
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        let properties = PropertyMapBuilder::new()
-            .insert("since", 2020i64)
-            .insert("weight", 0.85f64)
-            .build();
-
-        let temporal = BiTemporalInterval::new(
-            TimeRange::new(5000, 6000).unwrap(),
-            TimeRange::new(7000, 8000).unwrap(),
-        );
-
-        let operation = WalOperation::CreateEdge {
-            edge_id: EdgeId::new(99).unwrap(),
-            source: NodeId::new(1).unwrap(),
-            target: NodeId::new(2).unwrap(),
-            label: "FRIENDS_WITH".to_string(),
-            properties: properties.clone(),
-            temporal,
-        };
-
-        wal.append(operation)?;
-        wal.flush()?;
-
-        let entries = wal.read_from(LSN::initial())?;
-        assert_eq!(entries.len(), 1);
-
-        match &entries[0].operation {
-            WalOperation::CreateEdge {
-                edge_id,
-                source,
-                target,
-                label,
-                properties: p,
-                temporal: t,
-            } => {
-                assert_eq!(edge_id.as_u64(), 99);
-                assert_eq!(source.as_u64(), 1);
-                assert_eq!(target.as_u64(), 2);
-                assert_eq!(label, "FRIENDS_WITH");
-
-                assert_eq!(p.get("since").and_then(|v| v.as_int()), Some(2020));
-
-                assert_eq!(t.valid_time().start(), 5000);
-                assert_eq!(t.valid_time().end(), 6000);
-                assert_eq!(t.transaction_time().start(), 7000);
-                assert_eq!(t.transaction_time().end(), 8000);
-            }
-            _ => panic!("Expected CreateEdge operation"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_update_node_full_roundtrip() -> Result<()> {
-        use crate::core::property::PropertyMapBuilder;
-        use crate::core::temporal::TimeRange;
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        let properties = PropertyMapBuilder::new()
-            .insert("age", 31i64) // Updated age
-            .build();
-
-        let temporal = BiTemporalInterval::new(
-            TimeRange::new(10000, 20000).unwrap(),
-            TimeRange::from(15000),
-        );
-
-        let operation = WalOperation::UpdateNode {
-            node_id: NodeId::new(42).unwrap(),
-            version_id: VersionId::new(2).unwrap(),
-            label: "Person".to_string(),
-            properties: properties.clone(),
-            temporal,
-        };
-
-        wal.append(operation)?;
-        wal.flush()?;
-
-        let entries = wal.read_from(LSN::initial())?;
-        assert_eq!(entries.len(), 1);
-
-        match &entries[0].operation {
-            WalOperation::UpdateNode {
-                node_id,
-                version_id,
-                label,
-                properties: p,
-                temporal: t,
-            } => {
-                assert_eq!(node_id.as_u64(), 42);
-                assert_eq!(version_id.as_u64(), 2);
-                assert_eq!(label, "Person");
-                assert_eq!(p.get("age").and_then(|v| v.as_int()), Some(31));
-                assert_eq!(t.valid_time().start(), 10000);
-            }
-            _ => panic!("Expected UpdateNode operation"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_delete_with_temporal() -> Result<()> {
-        use crate::core::temporal::TimeRange;
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        let temporal = BiTemporalInterval::new(
-            TimeRange::new(100, 200).unwrap(),
-            TimeRange::new(300, 400).unwrap(),
-        );
-
-        wal.append(WalOperation::DeleteNode {
-            node_id: NodeId::new(5).unwrap(),
-            temporal,
-        })?;
-
-        wal.append(WalOperation::DeleteEdge {
-            edge_id: EdgeId::new(10).unwrap(),
-            temporal,
-        })?;
-
-        wal.flush()?;
-
-        let entries = wal.read_from(LSN::initial())?;
-        assert_eq!(entries.len(), 2);
-
-        // Verify DeleteNode temporal is preserved
-        match &entries[0].operation {
-            WalOperation::DeleteNode {
-                node_id,
-                temporal: t,
-            } => {
-                assert_eq!(node_id.as_u64(), 5);
-                assert_eq!(t.valid_time().start(), 100);
-                assert_eq!(t.valid_time().end(), 200);
-                assert_eq!(t.transaction_time().start(), 300);
-                assert_eq!(t.transaction_time().end(), 400);
-            }
-            _ => panic!("Expected DeleteNode"),
-        }
-
-        // Verify DeleteEdge temporal is preserved
-        match &entries[1].operation {
-            WalOperation::DeleteEdge {
-                edge_id,
-                temporal: t,
-            } => {
-                assert_eq!(edge_id.as_u64(), 10);
-                assert_eq!(t.valid_time().start(), 100);
-                assert_eq!(t.transaction_time().end(), 400);
-            }
-            _ => panic!("Expected DeleteEdge"),
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_segment_header() -> Result<()> {
-        use std::io::Read;
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Append an entry to create a segment
-        wal.append(WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: "Test".to_string(),
-            properties: PropertyMap::new(),
-            temporal: BiTemporalInterval::current(time::now()),
-        })?;
-        wal.flush()?;
-
-        // Read the segment file directly to verify header
-        let segment_path = temp_dir.path().join("000001.log");
-        let mut file = std::fs::File::open(&segment_path).unwrap();
-        let mut header = [0u8; 5];
-        file.read_exact(&mut header).unwrap();
-
-        // Verify magic bytes "GWAL"
-        assert_eq!(&header[0..4], b"GWAL");
-        // Verify version
-        assert_eq!(header[4], 1); // WAL_VERSION
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_recovery_preserves_all_data() -> Result<()> {
-        use crate::core::property::PropertyMapBuilder;
-        use crate::core::temporal::TimeRange;
-
-        let temp_dir = TempDir::new().unwrap();
-        let wal_dir = temp_dir.path().to_path_buf();
-
-        // Create data in WAL and simulate crash by dropping
-        let original_properties;
-        let original_temporal;
-        {
-            let config = WalConfig {
-                wal_dir: wal_dir.clone(),
-                ..Default::default()
-            };
-            let mut wal = WriteAheadLog::new(config)?;
-
-            original_properties = PropertyMapBuilder::new()
-                .insert("key", "value")
-                .insert("count", 42i64)
-                .build();
-            original_temporal = BiTemporalInterval::new(
-                TimeRange::new(1000, 2000).unwrap(),
-                TimeRange::new(3000, 4000).unwrap(),
-            );
-
-            wal.append(WalOperation::CreateNode {
-                node_id: NodeId::new(100).unwrap(),
-                label: "TestNode".to_string(),
-                properties: original_properties.clone(),
-                temporal: original_temporal,
-            })?;
-            wal.flush()?;
-            // WAL dropped here
-        }
-
-        // Recover from WAL
-        {
-            let config = WalConfig {
-                wal_dir: wal_dir.clone(),
-                ..Default::default()
-            };
-            let wal = WriteAheadLog::new(config)?;
-
-            let entries = wal.read_from(LSN::initial())?;
-            assert_eq!(entries.len(), 1);
-
-            match &entries[0].operation {
-                WalOperation::CreateNode {
-                    node_id,
-                    label,
-                    properties,
-                    temporal,
-                } => {
-                    assert_eq!(node_id.as_u64(), 100);
-                    assert_eq!(label, "TestNode");
-
-                    // Verify properties preserved
-                    assert_eq!(
-                        properties.get("key").and_then(|v| v.as_str()),
-                        Some("value")
-                    );
-                    assert_eq!(properties.get("count").and_then(|v| v.as_int()), Some(42));
-
-                    // Verify temporal preserved
-                    assert_eq!(temporal.valid_time().start(), 1000);
-                    assert_eq!(temporal.valid_time().end(), 2000);
-                    assert_eq!(temporal.transaction_time().start(), 3000);
-                    assert_eq!(temporal.transaction_time().end(), 4000);
-                }
-                _ => panic!("Expected CreateNode"),
-            }
-        }
-
-        Ok(())
-    }
-
-    // ========== Error Handling Tests for Coverage ==========
-
-    #[test]
-    fn test_wal_invalid_node_id_dos_protection() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let segment_path = temp_dir.path().join("000001.log");
-
-        // Manually craft WAL entry with invalid node_id (exceeds MAX_VALID_ID)
-        let mut buffer = Vec::new();
-
-        // Write header
-        buffer.extend_from_slice(&WAL_MAGIC);
-        buffer.push(WAL_VERSION);
-
-        // Write LSN
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-
-        // Write timestamp
-        buffer.extend_from_slice(&time::now().to_le_bytes());
-
-        // Write checksum placeholder
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // Write OpType::CreateNode (1)
-        buffer.push(1u8);
-
-        // Write INVALID node_id (exceeds MAX_VALID_ID)
-        let invalid_id = u64::MAX;
-        buffer.extend_from_slice(&invalid_id.to_le_bytes());
-
-        // Write minimal rest of entry
-        let label = "Test";
-        buffer.extend_from_slice(&(label.len() as u32).to_le_bytes());
-        buffer.extend_from_slice(label.as_bytes());
-
-        // Write empty properties
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // Write temporal data
-        let temporal = BiTemporalInterval::current(time::now());
-        temporal.serialize_into(&mut buffer);
-
-        std::fs::write(&segment_path, buffer)?;
-
-        // Attempt to read via WAL
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let wal = WriteAheadLog::new(config)?;
-        let result = wal.read_from(LSN::initial());
-
-        // Should fail with corrupted data error for invalid ID
-        assert!(result.is_err());
-        assert!(
-            matches!(
-                result.unwrap_err(),
-                crate::utils::error::Error::Storage(
-                    crate::utils::error::StorageError::CorruptedData(_)
-                )
-            ),
-            "Expected CorruptedData error for invalid node ID"
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_invalid_edge_id_dos_protection() -> Result<()> {
-        use crate::core::id::MAX_VALID_ID;
-
-        let temp_dir = TempDir::new().unwrap();
-        let segment_path = temp_dir.path().join("000001.log");
-
-        let mut buffer = Vec::new();
-        buffer.extend_from_slice(&WAL_MAGIC);
-        buffer.push(WAL_VERSION);
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-        buffer.extend_from_slice(&time::now().to_le_bytes());
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // OpType::CreateEdge (2)
-        buffer.push(2u8);
-
-        // Invalid edge_id
-        buffer.extend_from_slice(&(MAX_VALID_ID + 1).to_le_bytes());
-
-        // source and target (need to add label, properties, temporal for complete entry)
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-        buffer.extend_from_slice(&2u64.to_le_bytes());
-
-        // Add label
-        let label = "Test";
-        buffer.extend_from_slice(&(label.len() as u32).to_le_bytes());
-        buffer.extend_from_slice(label.as_bytes());
-
-        // Add empty properties
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // Add temporal
-        let temporal = BiTemporalInterval::current(time::now());
-        temporal.serialize_into(&mut buffer);
-
-        std::fs::write(&segment_path, buffer)?;
-
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let wal = WriteAheadLog::new(config)?;
-        let result = wal.read_from(LSN::initial());
-
-        // Should fail with corrupted data error for invalid edge ID
-        assert!(result.is_err());
-        assert!(
-            matches!(
-                result.unwrap_err(),
-                crate::utils::error::Error::Storage(
-                    crate::utils::error::StorageError::CorruptedData(_)
-                )
-            ),
-            "Expected CorruptedData error for invalid edge ID"
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_invalid_version_id_rejection() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let segment_path = temp_dir.path().join("000001.log");
-
-        let mut buffer = Vec::new();
-        buffer.extend_from_slice(&WAL_MAGIC);
-        buffer.push(WAL_VERSION);
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-        buffer.extend_from_slice(&time::now().to_le_bytes());
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // OpType::UpdateNode (3)
-        buffer.push(3u8);
-
-        // Valid node_id
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-
-        // Invalid version_id
-        buffer.extend_from_slice(&u64::MAX.to_le_bytes());
-
-        std::fs::write(&segment_path, buffer)?;
-
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let wal = WriteAheadLog::new(config)?;
-        let result = wal.read_from(LSN::initial());
-
-        assert!(result.is_err());
-        assert!(
-            matches!(
-                result.unwrap_err(),
-                crate::utils::error::Error::Storage(
-                    crate::utils::error::StorageError::CorruptedData(_)
-                )
-            ),
-            "Expected CorruptedData error for invalid version ID"
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_unknown_operation_type() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let segment_path = temp_dir.path().join("000001.log");
-
-        let mut buffer = Vec::new();
-        buffer.extend_from_slice(&WAL_MAGIC);
-        buffer.push(WAL_VERSION);
-        buffer.extend_from_slice(&1u64.to_le_bytes());
-        buffer.extend_from_slice(&time::now().to_le_bytes());
-        buffer.extend_from_slice(&0u32.to_le_bytes());
-
-        // UNKNOWN OpType (99)
-        buffer.push(99u8);
-
-        std::fs::write(&segment_path, buffer)?;
-
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let wal = WriteAheadLog::new(config)?;
-        let result = wal.read_from(LSN::initial());
-
-        assert!(result.is_err());
-        assert!(
-            matches!(
-                result.unwrap_err(),
-                crate::utils::error::Error::Storage(
-                    crate::utils::error::StorageError::CorruptedData(_)
-                )
-            ),
-            "Expected CorruptedData error for unknown operation type"
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_truncated_entry_graceful_stop() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Write 5 complete entries
-        for i in 1..=5 {
-            wal.append(WalOperation::CreateNode {
-                node_id: NodeId::new(i).unwrap(),
-                label: "Test".to_string(),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            })?;
-        }
-        wal.flush()?;
-        drop(wal);
-
-        // Truncate the file mid-entry (remove last 20 bytes)
-        let segment_path = temp_dir.path().join("000001.log");
-        let mut data = std::fs::read(&segment_path)?;
-        let truncate_len = data.len().saturating_sub(20);
-        data.truncate(truncate_len);
-        std::fs::write(&segment_path, data)?;
-
-        // Read should either error or stop gracefully at truncation
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let wal = WriteAheadLog::new(config)?;
-        let result = wal.read_from(LSN::initial());
-
-        // Accept either error (strict) or partial recovery (graceful)
-        match result {
-            Err(e) => {
-                // Truncation detected - this is valid behavior
-                assert!(
-                    e.to_string().contains("Buffer too short") || e.to_string().contains("corrupt")
-                );
-            }
-            Ok(entries) => {
-                // Graceful recovery - should have at least some entries
-                assert!(entries.len() >= 4 && entries.len() <= 5);
-            }
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_corrupted_checksum_detection() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-        wal.append(WalOperation::CreateNode {
-            node_id: NodeId::new(1).unwrap(),
-            label: "Test".to_string(),
-            properties: PropertyMap::new(),
-            temporal: BiTemporalInterval::current(time::now()),
-        })?;
-        wal.flush()?;
-        drop(wal);
-
-        // Corrupt some data (not just checksum) to trigger corruption detection
-        let segment_path = temp_dir.path().join("000001.log");
-        let mut data = std::fs::read(&segment_path)?;
-
-        // Corrupt data in the middle (operation type byte after header)
-        if data.len() > WAL_HEADER_SIZE + 20 {
-            // Corrupt the operation section
-            data[WAL_HEADER_SIZE + 16] ^= 0xFF;
-            data[WAL_HEADER_SIZE + 17] ^= 0xFF;
-            std::fs::write(&segment_path, data)?;
-        }
-
-        // Attempt to read - should detect corruption or fail gracefully
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let wal = WriteAheadLog::new(config)?;
-        let result = wal.read_from(LSN::initial());
-
-        // Corrupted data should either error or be detected as invalid
-        // This test documents the current behavior
-        match result {
-            Err(_) => {} // Expected - corruption detected
-            Ok(entries) => {
-                // If implementation doesn't validate checksums yet, might succeed
-                // but this test documents that corruption detection is desired
-                assert!(entries.len() <= 1);
-            }
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_wal_segment_cleanup_respects_retention() -> Result<()> {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Create config with very small segment size to force rotation
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            segment_size: 256, // Very small to force rotation
-            segments_to_retain: 5,
-            durability_mode: DurabilityMode::Synchronous,
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Create enough entries to force multiple segment rotations
-        for i in 0..150 {
-            wal.append(WalOperation::CreateNode {
-                node_id: NodeId::new(i as u64 + 1).unwrap(),
-                label: format!("TestNode{}", i),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            })?;
-
-            // Flush every few entries to force rotations
-            if i % 5 == 0 {
-                wal.flush()?;
-            }
-        }
-
-        wal.flush()?;
-
-        // Trigger cleanup
-        wal.cleanup_old_segments()?;
-
-        // Count remaining segment files
-        let segment_count = std::fs::read_dir(temp_dir.path())?
-            .filter_map(|e| e.ok())
-            .filter(|e| {
-                e.path()
-                    .extension()
-                    .and_then(|s| s.to_str())
-                    .map(|s| s == "log")
-                    .unwrap_or(false)
-            })
-            .count();
-
-        // Should have at most segments_to_retain + current segment
-        assert!(
-            segment_count <= 6,
-            "Expected ≤6 segments (5 retained + 1 current), found {}",
-            segment_count
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_async_mode_integration() -> Result<()> {
-        use std::sync::{Arc, Mutex};
-        use std::thread;
-        use std::time::Duration;
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            durability_mode: DurabilityMode::async_mode_validated(50)?, // 50ms flush interval
-            ..Default::default()
-        };
-
-        let wal = Arc::new(Mutex::new(WriteAheadLog::new(config)?));
-
-        // Start background flush thread (initializes AsyncWalWriter)
-        WriteAheadLog::start_flush_thread(Arc::clone(&wal));
-
-        // Verify AsyncWalWriter was created
-        {
-            let wal_guard = wal.lock().unwrap();
-            assert!(
-                wal_guard.async_writer.is_some(),
-                "AsyncWalWriter should be initialized for Async mode"
-            );
-        }
-
-        // Append multiple entries (should return immediately without blocking on fsync)
-        for i in 0..10 {
-            let mut wal_guard = wal.lock().unwrap();
-            wal_guard.append(WalOperation::CreateNode {
-                node_id: NodeId::new(i as u64 + 1).unwrap(),
-                label: format!("AsyncNode{}", i),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            })?;
-        }
-
-        // Wait for background sync (flush interval + buffer time)
-        thread::sleep(Duration::from_millis(150));
-
-        // Verify entries were synced
-        let entries = {
-            let wal_guard = wal.lock().unwrap();
-            let segment_path = wal_guard.segment_path(1);
-            wal_guard.read_segment(&segment_path, LSN::initial())?
-        };
-
-        assert_eq!(entries.len(), 10, "All entries should be synced to disk");
-
-        // Verify LSNs are sequential
-        for (i, entry) in entries.iter().enumerate() {
-            assert_eq!(entry.lsn.0, (i + 1) as u64);
-        }
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_async_mode_high_throughput() -> Result<()> {
-        use std::sync::{Arc, Mutex};
-        use std::thread;
-        use std::time::{Duration, Instant};
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            durability_mode: DurabilityMode::async_mode_validated(100)?, // 100ms flush interval
-            ..Default::default()
-        };
-
-        let wal = Arc::new(Mutex::new(WriteAheadLog::new(config)?));
-        WriteAheadLog::start_flush_thread(Arc::clone(&wal));
-
-        // Measure throughput
-        let start = Instant::now();
-        let entry_count = 1000;
-
-        for i in 0..entry_count {
-            let mut wal_guard = wal.lock().unwrap();
-            wal_guard.append(WalOperation::CreateNode {
-                node_id: NodeId::new(i as u64 + 1).unwrap(),
-                label: format!("HighThroughputNode{}", i),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            })?;
-        }
-
-        let elapsed = start.elapsed();
-        let throughput = entry_count as f64 / elapsed.as_secs_f64();
-
-        // Async mode should achieve reasonable throughput even on CI
-        // Lowered from 10k to 3k to account for slower/shared CI runners (especially Windows)
-        // This still validates that async mode is significantly faster than sync mode
-        assert!(
-            throughput > 3_000.0,
-            "Expected throughput >3k ops/sec, got {:.0}",
-            throughput
-        );
-
-        // Wait for final sync
-        thread::sleep(Duration::from_millis(200));
-
-        // Verify all entries are durable
-        let entries = {
-            let wal_guard = wal.lock().unwrap();
-            let segment_path = wal_guard.segment_path(1);
-            wal_guard.read_segment(&segment_path, LSN::initial())?
-        };
-
-        assert_eq!(
-            entries.len(),
-            entry_count,
-            "All entries should be synced to disk"
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_async_mode_graceful_shutdown() -> Result<()> {
-        use std::sync::{Arc, Mutex};
-        use std::thread;
-        use std::time::Duration;
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            durability_mode: DurabilityMode::async_mode_validated(100)?,
-            ..Default::default()
-        };
-
-        let wal = Arc::new(Mutex::new(WriteAheadLog::new(config)?));
-        WriteAheadLog::start_flush_thread(Arc::clone(&wal));
-
-        // Append some entries
-        for i in 0..5 {
-            let mut wal_guard = wal.lock().unwrap();
-            wal_guard.append(WalOperation::CreateNode {
-                node_id: NodeId::new(i as u64 + 1).unwrap(),
-                label: format!("ShutdownNode{}", i),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            })?;
-        }
-
-        // Drop WAL (should trigger graceful shutdown of AsyncWalWriter)
-        drop(wal);
-
-        // Give background thread time to finish
-        thread::sleep(Duration::from_millis(50));
-
-        // Verify entries were synced before shutdown
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            durability_mode: DurabilityMode::Synchronous,
-            ..Default::default()
-        };
-
-        let wal = WriteAheadLog::new(config)?;
-        let segment_path = wal.segment_path(1);
-        let entries = wal.read_segment(&segment_path, LSN::initial())?;
-
-        assert_eq!(
-            entries.len(),
-            5,
-            "All entries should be synced before shutdown"
-        );
-
-        Ok(())
-    }
-
-    /// Test buffer reuse correctness for Issue #20
-    ///
-    /// Verifies that reusing the serialization buffer doesn't cause data corruption
-    /// or cross-contamination between sequential WAL entries.
-    #[test]
-    fn test_serialize_buffer_reuse_correctness() -> Result<()> {
-        use crate::core::property::PropertyMapBuilder;
-
-        let temp_dir = TempDir::new().unwrap();
-        let config = WalConfig {
-            wal_dir: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-
-        let mut wal = WriteAheadLog::new(config)?;
-
-        // Create a series of operations with different data sizes
-        // to ensure buffer resizing works correctly
-        let operations = vec![
-            // Small entry
-            WalOperation::CreateNode {
-                node_id: NodeId::new(1).unwrap(),
-                label: "A".to_string(),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            },
-            // Larger entry with properties
-            WalOperation::CreateNode {
-                node_id: NodeId::new(2).unwrap(),
-                label: "LongerLabel".to_string(),
-                properties: PropertyMapBuilder::new()
-                    .insert("key1", "value1")
-                    .insert("key2", 42i64)
-                    .insert("key3", true)
-                    .build(),
-                temporal: BiTemporalInterval::current(time::now()),
-            },
-            // Small entry again (tests buffer shrinking/reuse)
-            WalOperation::CreateNode {
-                node_id: NodeId::new(3).unwrap(),
-                label: "B".to_string(),
-                properties: PropertyMap::new(),
-                temporal: BiTemporalInterval::current(time::now()),
-            },
-            // Different operation type
-            WalOperation::CreateEdge {
-                edge_id: EdgeId::new(1).unwrap(),
-                source: NodeId::new(1).unwrap(),
-                target: NodeId::new(2).unwrap(),
-                label: "KNOWS".to_string(),
-                properties: PropertyMapBuilder::new().insert("weight", 0.95f64).build(),
-                temporal: BiTemporalInterval::current(time::now()),
-            },
-        ];
-
-        // Append all operations (buffer reuse happens here)
-        for op in operations {
-            wal.append(op)?;
-        }
-
-        wal.flush()?;
-
-        // Read back and verify all data is correct
-        let entries = wal.read_from(LSN::initial())?;
-        assert_eq!(entries.len(), 4, "Should read back all 4 entries");
-
-        // Verify first entry (small)
-        match &entries[0].operation {
-            WalOperation::CreateNode { node_id, label, .. } => {
-                assert_eq!(node_id.as_u64(), 1);
-                assert_eq!(label, "A");
-            }
-            _ => panic!("Expected CreateNode"),
-        }
-
-        // Verify second entry (large with properties)
-        match &entries[1].operation {
-            WalOperation::CreateNode {
-                node_id,
-                label,
-                properties,
-                ..
-            } => {
-                assert_eq!(node_id.as_u64(), 2);
-                assert_eq!(label, "LongerLabel");
-                assert_eq!(
-                    properties.get("key1").and_then(|v| v.as_str()),
-                    Some("value1")
-                );
-                assert_eq!(properties.get("key2").and_then(|v| v.as_int()), Some(42));
-                assert_eq!(properties.get("key3").and_then(|v| v.as_bool()), Some(true));
-            }
-            _ => panic!("Expected CreateNode"),
-        }
-
-        // Verify third entry (small again - tests buffer didn't retain old data)
-        match &entries[2].operation {
-            WalOperation::CreateNode {
-                node_id,
-                label,
-                properties,
-                ..
-            } => {
-                assert_eq!(node_id.as_u64(), 3);
-                assert_eq!(label, "B");
-                // Should be empty, not contaminated from previous large entry
-                assert_eq!(properties.len(), 0);
-            }
-            _ => panic!("Expected CreateNode"),
-        }
-
-        // Verify fourth entry (different operation type)
-        match &entries[3].operation {
-            WalOperation::CreateEdge {
-                edge_id,
-                source,
-                target,
-                properties,
-                ..
-            } => {
-                assert_eq!(edge_id.as_u64(), 1);
-                assert_eq!(source.as_u64(), 1);
-                assert_eq!(target.as_u64(), 2);
-                assert_eq!(
-                    properties.get("weight").and_then(|v| v.as_float()),
-                    Some(0.95)
-                );
-            }
-            _ => panic!("Expected CreateEdge"),
-        }
-
-        Ok(())
-    }
 }
