+++ src/storage/wal_reader.rs
@@ -5,7 +5,8 @@
 
 use std::path::Path;
 
-use super::wal::{DurabilityMode, LSN, WalConfig, WalEntry, WriteAheadLog};
+use super::wal::segment_reader;
+use super::wal::{LSN, WalEntry};
 use crate::utils::error::Result;
 
 /// Read WAL entries from a directory, starting from the specified LSN.
@@ -23,16 +24,5 @@ use crate::utils::error::Result;
 ///
 /// A vector of WAL entries sorted by LSN.
 pub fn read_wal_entries(wal_dir: &Path, start_lsn: LSN) -> Result<Vec<WalEntry>> {
-    // Create a minimal config just for reading
-    let config = WalConfig {
-        wal_dir: wal_dir.to_path_buf(),
-        segment_size: 64 * 1024 * 1024, // Default, not used for reading
-        segments_to_retain: 10,         // Default, not used for reading
-        durability_mode: DurabilityMode::Synchronous, // Not used for reading
-    };
-
-    // Create a minimal WAL instance for reading
-    // Note: WriteAheadLog::new() doesn't create segment files, just initializes the struct
-    let wal = WriteAheadLog::new(config)?;
-    wal.read_from(start_lsn)
+    segment_reader::read_entries_from_dir(wal_dir, start_lsn)
 }
